refactor: 카드 브랜드 관련 로직 유틸 함수로 분리
